Continue

+++ scripts/run-docker-container-2.py
@@ -10,18 +10,13 @@ def parse_arguments():
     """
     parser = argparse.ArgumentParser(description='Run course Docker container.')
     parser.add_argument('-n', '--image_name', type=str, help='Docker image.', required=True)
-    parser.add_argument('-c', '--container_name', type=str, help="Docker container name.", required=True)
     parser.add_argument('-p', '--port', type=int, help='Port number.', required=True)
     parser.add_argument('-v', '--volume_string', type=str, help="Volumes to be mounted to /work.", required=True)
-    parser.add_argument('-i', '--interactive_mode', type=bool, help="If True, run with -it flag.")
+    parser.add_argument('-i', '--interactive_mode', type=str, help="If True, run with -it flag.", required=True)
     return parser.parse_args()
 
 def parse_n(arg_n):
     return(arg_n.startswith('dreg.cloud.sdu.dk/ucloud-courses/'))
-
-def parse_c(arg_c):
-    # TODO
-    pass
 
 def parse_p(arg_p):
     return(type(arg_p) == int)
@@ -40,29 +35,54 @@ def parse_v(arg_v):
     
     # Check and return results 
     if all(isinstance(v, bool) for v in v_exists) and all(v_exists):
-        return(True)
+        return(True, v_list)
     else:
         print("ERROR: The following directories specified in VOLUMES do not exist:")
         for i in range(0, len(v_exists)):
             if not v_exists[i]:
                 print("* %s"%(v_list[i]))
-        return(False)
+        return(False, [])
 
 def parse_i(arg_i):
-    return(type(arg_i) == bool)
+    if arg_i == "True" or arg_i == "False":
+        val = True if arg_i == "True" else False
+        return(True, val)
+    return(False, None)
 
 if __name__ == "__main__":
-
+    # Parse the arguments 
     args = parse_arguments()
-    print(parse_n(args.image_name))
-    print(parse_p(args.port))
-    print(parse_i(args.interactive_mode))
-    print(parse_v(args.volume_string))
+    i_isvalid, i_value = parse_i(args.interactive_mode)
+    v_list_bool, v_list_parsed = parse_v(args.volume_string)
 
-    call_string = "docker run --rm "
-    call_string = call_string + "%s"%('-it ' if parse_i(args.interactive_mode) else '')
-    # TODO: Finalize call_string
+    # Get results from parser functions
+    res = {'image_name': parse_n(args.image_name), 'port': parse_p(args.port), 'volume_string': v_list_bool, 'interactive_mode': i_isvalid}
 
-    
-    # Parts the args to form an argument string for `docker run`
-    os.system('docker run --rm dreg.cloud.sdu.dk/ucloud-courses/sdu-test__1234:2024-01-01')
+    # Error handling from parser functions
+    if any(value == False for value in res.values()):
+        if not res['image_name']: 
+            print("ERROR: Invalid IMAGE_NAME.")
+        if not res['port']:
+            print("ERROR: Invalid PORT. Must be an integer.")
+        if not res["volume_string"]:
+            print("ERROR: Invalid VOLUMES.")
+        if not res["interactive_mode"]:
+            print("ERROR: Invalid value for INTERACTIVE_MODE. Must be 'True' or 'False'")
+        exit(1)
+    else: 
+        # Create call string for `docker run` 
+        call_string = "docker run --rm "
+        call_string += "%s "%('-it' if i_value else '')
+        call_string += "%s %s:%s "%('-p', args.port, args.port)
+        
+        for v in v_list_parsed:
+            call_string += "%s %s:%s "%('-v', v, "/work/" + os.path.basename(v))
+        
+        call_string += " %s"%(args.image_name)
+        
+        try:
+            # Start container 
+            print("\nStarting docker container (in %s mode).\n"%("interactive" if i_value else "non-interactive"))
+            os.system(call_string)
+        except OSError as e:
+            exit(e)
